+++ src/Network/Network/Common/NetworkResourceManagerProfile.cs
@@ -32,7 +32,7 @@ namespace Microsoft.Azure.Commands.Network
         {
             get
             {
-                lock(_lock)
+                lock (_lock)
                 {
                     if (_mapper == null)
                     {
@@ -86,6 +86,37 @@ namespace Microsoft.Azure.Commands.Network
 
                     cnmProp.SetValue(cnmObj, list);
                 }
+            }
+        }
+
+        public class VirtualRouterConverter : ITypeConverter<MNM.VirtualRouter, CNM.PSVirtualRouter>
+        {
+            public CNM.PSVirtualRouter Convert(MNM.VirtualRouter source, CNM.PSVirtualRouter destination, ResolutionContext context)
+            {
+                var router = new CNM.PSVirtualRouter()
+                {
+                    Id = source.Id,
+                    Name = source.Name,
+                    Etag = source.Etag,
+                    Location = source.Location,
+                    Type = source.Type,
+                    ProvisioningState = source.ProvisioningState,
+                    VirtualRouterAsn = (uint) source.VirtualRouterAsn
+                };
+                if (source.HostedGateway != null)
+                {
+                    router.HostedGateway = new CNM.PSResourceId()
+                    {
+                        Id = source.HostedGateway.Id
+                    };
+                }
+                if (source.VirtualRouterIps != null)
+                {
+                    router.VirtualRouterIps = source.VirtualRouterIps.ToList<string>();
+                }
+
+                router.Peerings = new List<CNM.PSVirtualRouterPeer>();
+                return router;
             }
         }
 
@@ -220,7 +251,8 @@ namespace Microsoft.Azure.Commands.Network
 
         private static void Initialize()
         {
-            var config = new MapperConfiguration(cfg => {
+            var config = new MapperConfiguration(cfg =>
+            {
                 cfg.AddProfile<NetworkResourceManagerProfile>();
                 cfg.CreateMap<CNM.PSResourceId, MNM.SubResource>();
                 cfg.CreateMap<MNM.SubResource, CNM.PSResourceId>();
@@ -266,7 +298,7 @@ namespace Microsoft.Azure.Commands.Network
                 cfg.CreateMap<CNM.PSVirtualNetworkBgpCommunities, MNM.VirtualNetworkBgpCommunities>();
                 cfg.CreateMap<CNM.PSSubnet, MNM.Subnet>()
                     .ForMember(dest => dest.AddressPrefix, opt => opt.Ignore())
-                    .ForMember(dest=> dest.AddressPrefixes, opt => opt.Ignore())
+                    .ForMember(dest => dest.AddressPrefixes, opt => opt.Ignore())
                     .AfterMap((src, dest) =>
                     {
                         if (GeneralUtilities.HasMoreThanOneElement(src.AddressPrefix))
@@ -295,9 +327,9 @@ namespace Microsoft.Azure.Commands.Network
                         {
                             dest.AddressPrefix = src.AddressPrefixes?.ToList();
                         }
-                        else if(!string.IsNullOrEmpty(src.AddressPrefix))
+                        else if (!string.IsNullOrEmpty(src.AddressPrefix))
                         {
-                            dest.AddressPrefix = new List<string> {src.AddressPrefix};
+                            dest.AddressPrefix = new List<string> { src.AddressPrefix };
                         }
                     });
                 cfg.CreateMap<MNM.IPConfiguration, CNM.PSIPConfiguration>();
@@ -312,7 +344,7 @@ namespace Microsoft.Azure.Commands.Network
 
                 // MNM to CNM
                 cfg.CreateMap<MNM.IPAddressAvailabilityResult, CNM.PSIPAddressAvailabilityResult>();
-                
+
                 // Avaliable endpoint services
                 // CNM to MNM
                 cfg.CreateMap<CNM.PSEndpointServiceResult, MNM.EndpointServiceResult>();
@@ -503,7 +535,7 @@ namespace Microsoft.Azure.Commands.Network
                 cfg.CreateMap<MNM.ConnectivityInformation, CNM.PSConnectivityInformation>();
                 cfg.CreateMap<MNM.ConnectivityHop, CNM.PSConnectivityHop>();
                 cfg.CreateMap<MNM.ConnectivityIssue, CNM.PSConnectivityIssue>();
-                
+
                 // AvailableProvidersList
                 // CNM to MNM
                 cfg.CreateMap<CNM.PSAvailableProvidersList, MNM.AvailableProvidersList>();
@@ -677,7 +709,7 @@ namespace Microsoft.Azure.Commands.Network
                     {
                         MapSecurityRuleCommandToManagement<CNM.PSEffectiveSecurityRule, MNM.EffectiveNetworkSecurityRule>(src, dest);
                      });
-                
+
                 // MNM to CNM
                 cfg.CreateMap<MNM.EffectiveNetworkSecurityGroup, CNM.PSEffectiveNetworkSecurityGroup>();
                 cfg.CreateMap<MNM.EffectiveNetworkSecurityGroupAssociation, CNM.PSEffectiveNetworkSecurityGroupAssociation>();
@@ -689,7 +721,7 @@ namespace Microsoft.Azure.Commands.Network
 
                 // ExpressRoutePortsLocation
                 // CNM to MNM
-                cfg.CreateMap<CNM.PSExpressRoutePortsLocation, MNM.ExpressRoutePortsLocation>();                
+                cfg.CreateMap<CNM.PSExpressRoutePortsLocation, MNM.ExpressRoutePortsLocation>();
                 cfg.CreateMap<CNM.PSExpressRoutePortsLocationBandwidths, MNM.ExpressRoutePortsLocationBandwidths>();
 
                 // MNM to CNM
@@ -843,7 +875,7 @@ namespace Microsoft.Azure.Commands.Network
                 // Application Gateways
                 // CNM to MNM
                 cfg.CreateMap<CNM.PSApplicationGateway, MNM.ApplicationGateway>();
-                cfg.CreateMap<CNM.PSApplicationGatewaySku, MNM.ApplicationGatewaySku>(); 
+                cfg.CreateMap<CNM.PSApplicationGatewaySku, MNM.ApplicationGatewaySku>();
                 cfg.CreateMap<CNM.PSApplicationGatewaySslPolicy, MNM.ApplicationGatewaySslPolicy>()
                     .AfterMap((src, dest) =>
                     {
@@ -1107,6 +1139,15 @@ namespace Microsoft.Azure.Commands.Network
                 // MNM to CNM
                 cfg.CreateMap<MNM.AzureFirewallFqdnTag, CNM.PSAzureFirewallFqdnTag>();
 
+                // Azure Firewall Policies
+                // CNM to MNM
+                cfg.CreateMap<CNM.PSAzureFirewallPolicyRuleCollectionGroup, MNM.FirewallPolicyRuleGroup>();
+                cfg.CreateMap<CNM.PSAzureFirewallPolicy, MNM.FirewallPolicy>();
+
+                // MNM to CNM
+                cfg.CreateMap<MNM.FirewallPolicyRuleGroup, CNM.PSAzureFirewallPolicyRuleCollectionGroup>();
+                cfg.CreateMap<MNM.FirewallPolicy, CNM.PSAzureFirewallPolicy>();
+
                 // Virtual Network Tap
                 // CNM to MNM
                 cfg.CreateMap<CNM.PSVirtualNetworkTap, MNM.VirtualNetworkTap>();
@@ -1150,6 +1191,15 @@ namespace Microsoft.Azure.Commands.Network
                 // MNM to CNM
                 cfg.CreateMap<MNM.BastionHost, CNM.PSBastion>();
                 cfg.CreateMap<MNM.BastionHostIPConfiguration, CNM.PSBastionIPConfiguration>();
+
+                // Virtual Router
+                // CNM to MNM
+                cfg.CreateMap<CNM.PSVirtualRouter, MNM.VirtualRouter>();
+                cfg.CreateMap<CNM.PSVirtualRouterPeer, MNM.VirtualRouterPeering>();
+
+                // MNM to CNM
+                cfg.CreateMap<MNM.VirtualRouter, CNM.PSVirtualRouter>().ConvertUsing(new VirtualRouterConverter());
+                cfg.CreateMap<MNM.VirtualRouterPeering, CNM.PSVirtualRouterPeer>();
             });
 
             _mapper = config.CreateMapper();