+++ src/Network/Network/ChangeLog.md
@@ -35,6 +35,7 @@
 * New Generations and SKUs for VirtualNetworkGateways
   - Introduce new Generations for VirtualNetworkGateways.
   - Introduce new high throughput SKUs for VirtualNetworkGateways.
+* New cmdlets to support the azure firewall policy
 * Add support for child resource RouteTables of VirtualHub
 	- New cmdlets added:
 		- Add-AzVirtualHubRoute